# Docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: BOSSNYUMBA Property Management API
  description: |
    Comprehensive REST API for the BOSSNYUMBA multi-tenant property management SaaS platform.
    
    ## Authentication
    All endpoints (except `/auth/login`, `/auth/register`, `/auth/forgot-password`, `/auth/demo-users`,
    and `/payments/mpesa/callback`) require a valid JWT Bearer token in the `Authorization` header.
    
    ## Multi-Tenancy
    All data is scoped to the authenticated user's tenant. The tenant context is derived from the JWT token.
    
    ## Pagination
    List endpoints support pagination via `page` and `pageSize` query parameters.
    Responses include a `pagination` object with `page`, `pageSize`, `totalItems`, `totalPages`,
    `hasNextPage`, and `hasPreviousPage` fields.
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable description"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: BOSSNYUMBA Engineering
    email: engineering@bossnyumba.com
  license:
    name: Proprietary
    url: https://bossnyumba.com/terms
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://api.bossnyumba.com/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /work-orders:
    get:
      tags: [Work Orders]
      summary: List work orders
      operationId: listWorkOrders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [SUBMITTED, TRIAGED, APPROVED, ASSIGNED, IN_PROGRESS, COMPLETED, CANCELLED]
        - name: priority
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, EMERGENCY]
        - name: category
          in: query
          schema:
            type: string
            enum: [PLUMBING, ELECTRICAL, HVAC, STRUCTURAL, OTHER]
        - name: propertyId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of work orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /work-orders/{id}:
    get:
      tags: [Work Orders]
      summary: Get work order by ID
      operationId: getWorkOrder
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Work order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /feedback:
    get:
      tags: [Feedback]
      summary: List feedback
      operationId: listFeedback
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: type
          in: query
          schema:
            type: string
            enum: [general, bug, feature, improvement]
      responses:
        '200':
          description: List of feedback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      tags: [Feedback]
      summary: Submit feedback
      operationId: submitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitFeedbackRequest'
      responses:
        '201':
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /feedback/{id}:
    get:
      tags: [Feedback]
      summary: Get feedback by ID
      operationId: getFeedback
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Feedback details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /feedback/complaints:
    post:
      tags: [Feedback]
      summary: Create a complaint
      operationId: createComplaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComplaintRequest'
      responses:
        '201':
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /feedback/complaints/{id}:
    get:
      tags: [Feedback]
      summary: Get complaint by ID
      operationId: getComplaint
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /feedback/complaints/{id}/resolve:
    put:
      tags: [Feedback]
      summary: Resolve a complaint
      operationId: resolveComplaint
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                resolutionNotes:
                  type: string
      responses:
        '200':
          description: Complaint resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

# =============================================================================
# COMPONENTS
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from POST /auth/login.
        Include as: `Authorization: Bearer <token>`
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource unique identifier
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Items per page
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean

    # =========================================================================
    # Auth schemas
    # =========================================================================
...
    SubmitFeedbackRequest:
      type: object
      required: [type, subject, message]
      properties:
        type:
          type: string
          enum: [general, bug, feature, improvement]
        subject:
          type: string
          maxLength: 200
        message:
          type: string
          maxLength: 5000
        rating:
          type: integer
          minimum: 1
          maximum: 5
        context:
          type: object
    CreateComplaintRequest:
      type: object
      required: [subject, description]
      properties:
        subject:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        category:
          type: string
          enum: [maintenance, neighbor, payment, lease, other]
        relatedEntityType:
          type: string
        relatedEntityId:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium

    # =========================================================================
    # Work Order schemas
    # =========================================================================
...